+++ CHANGES.md
@@ -2,7 +2,10 @@
 
 ### 0.1.1
 
-* Added `bijection-json`
+* Adds `bijection-json`
+* Adds `bijection-protobuf`
+* Adds `TEnumCodec` for conversion of `TEnum` -> `Int`
+* Change `biject` method in `Bijection` object to `asMethod` for clarity
 
 ### 0.1.0